requirements.txt
# Core Dependencies
faster-whisper==1.0.3
torch>=2.0.0
torchaudio>=2.0.0

# Audio Processing
soundfile==0.12.1
librosa==0.10.1
noisereduce==3.0.2

# Utilities
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
PyYAML>=6.0
